+++ src/components/Appbar.js
@@ -14,7 +14,6 @@ import { ThemeProvider } from '@mui/material';
 import Theme from '../Theme';
 import { useNavigate } from 'react-router';
 import { useTranslation } from 'react-i18next';
-import AnimatedIcon from './AnimatedIcon';
 import routes from '../routes.json';