+++ changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Animation when opening a new
 - New experiment on experimental page
+- Animated social icons
 
 ### Fixed
 - Date month on news